feat Services: included exception handling in the UserService and BillService, created NotFoundException and ControllerAdvisor, included BeanValidation and modified classes for better standardization.

+++ src/main/java/com/taxislibres/technicaltest/Services/UserService.java
@@ -1,9 +1,9 @@
 package com.taxislibres.technicaltest.Services;
 
+import com.taxislibres.technicaltest.Exceptions.NotFoundException;
 import com.taxislibres.technicaltest.Models.User;
 import com.taxislibres.technicaltest.Repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,19 +22,28 @@ public class UserService {
     }
 
     public Optional<User> getUserById(Long id){
-        return repository.findById(id);
+        var user = repository.findById(id);
+        if(user.isEmpty()) throw new NotFoundException("User", id);
+        return user;
     }
 
     public User createUser(User user){
         return repository.save(user);
     }
 
-    public User updateUser(User user, Long userId) {
-        user.setId(userId);
-        return repository.save(user);
+    public User updateUser(User newUser, Long userId) {
+        var user = repository.findById(userId);
+        if(user.isEmpty()) throw new NotFoundException("User", userId);
+        if(newUser.getName() == null) newUser.setName(user.get().getName());
+        if(newUser.getAge() == 0) newUser.setAge(user.get().getAge());
+        if(newUser.getEmail() == null) newUser.setEmail(user.get().getEmail());
+        newUser.setId(user.get().getId());
+        return repository.save(newUser);
     }
 
     public void deleteUser(Long userId){
-         repository.deleteById(userId);
+        var user = repository.findById(userId);
+        if(user.isEmpty()) throw new NotFoundException("User", userId);
+        repository.deleteById(userId);
     }
 }
